+++ MOD/MinidoracatUIFor42/Contents/mods/MinidoracatUIFor42/42/media/lua/client/MinidoracatUI/Widgets/FloatButton.lua
@@ -229,6 +229,7 @@ function FloatButton.new(opts)
     o.getTooltipText = opts.getTooltip
     o:initialise()
     o:addToUIManager()
+    o:setAlwaysOnTop(o.alwaysOnTop)
     return o
 end